Fixed Clashes. Moves should deavtivate properly now
The problem with the characters going up to infinity was that their states were never reaching the conditions needed to qualify for deactivating a move. When they got hit, they dont qualify for being able to move, locking them from a whole chunk of the Update function. TL;DR: Had to modify the conditions in DeactivateMove() to take into account StartupMove OR ActiveMove being -1, so that DeactivateMove can Properly set things back to zero.
Also, keep in mind the HangTime of certain Moves, just in case certain moves juggle by themselves.
I altered the fallmultiplier value a bit as well